Moved libtx into its own crate
