+++ libtx/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "grin_libtx"
+version = "0.4.2"
+authors = ["Grin Developers <mimblewimble@lists.launchpad.net>"]
+description = "Simple, private and scalable cryptocurrency implementation based on the MimbleWimble chain format."
+license = "Apache-2.0"
+repository = "https://github.com/mimblewimble/grin"
+keywords = [ "crypto", "grin", "mimblewimble" ]
+workspace = '..'
+publish = false
+
+[dependencies]
+blake2-rfc = "0.2"
+failure = "0.1"
+failure_derive = "0.1"
+log = "0.4"
+rand = "0.5"
+serde = "1"
+serde_derive = "1"
+uuid = { version = "0.6", features = ["serde", "v4"] }
+
+grin_core = { path = "../core", version = "0.4.2" }
+grin_keychain = { path = "../keychain", version = "0.4.2" }
+grin_util = { path = "../util", version = "0.4.2" }
