+++ libtx/src/error.rs
