+++ libtx/src/reward.rs
@@ -16,12 +16,13 @@
 //! reward.
 use keychain::{Identifier, Keychain};
 
+use aggsig;
+use proof;
 use core::consensus::reward;
 use core::core::transaction::kernel_sig_msg;
 use core::core::KernelFeatures;
 use core::core::{Output, OutputFeatures, TxKernel};
-use libtx::error::Error;
-use libtx::{aggsig, proof};
+use error::Error;
 use util::static_secp_instance;
 
 /// output a reward output
